requirements.txt
django
channels
django-bootstrap3
djangorestframework
django-bootstrap-form
django-tables2
gitpython
requests
pygithub
Celery
pycryptodome
asgi_redis
markdown
django-markdownx
django-contrib-comments
django-notifications-hq
django-allauth
requirements-parser
django-model-utils
waliki[all]
django-bootstrap-breadcrumbs
travispy
GitPython
coveralls
sphinx
sqlalchemy
